requirements.txt
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.30.0
six==1.16.0
sniffio==1.3.0
tornado==6.1
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
